==> text-mining/notebooks/classification/04_predict.ipynb <==
import joblib

# Install gdown to download from Google Drive
!pip install gdown

# Download the saved model from Google Drive
!gdown --id 1eHZe445OQNc2GE7rdzVc7i0-akJPcLuK -O bert-toxic-comment-classification.zip

# Unzip the ZIP file
import zipfile

# Extract the ZIP file into the specified directory
with zipfile.ZipFile("bert-toxic-comment-classification.zip", 'r') as zip_ref:
    zip_ref.extractall("bert-toxic-comment-classification")  # Extracts into the specified directory


# Download the saved model from Google Drive
!gdown --id 14j1FohHhUFuJKiX1PGq4P538H4ja_KjO -O logistic_regression_multilabel.zip

# Unzip the ZIP file into the current directory
import zipfile

with zipfile.ZipFile("logistic_regression_multilabel.zip", 'r') as zip_ref:
    zip_ref.extractall()  # Extracts into the current working directory

print("Files successfully extracted into the current directory.")


# Download the ZIP file from Google Drive TODO
!gdown --id 1e4eT92P-i_QGcBEwHKf66iLn1aSpMr8e -O tokenized_test_data.zip

# Unzip the file
import shutil
shutil.unpack_archive('tokenized_test_data.zip')

print("File unzipped. Ready to load.")


# Download the saved model from Google Drive
!gdown --id 1LDmkBpIjEJy3rMghZEF3evM2Imt0wu9g -O test_processed.zip

# Unzip the ZIP file
with zipfile.ZipFile("test_processed.zip", 'r') as zip_ref:
    zip_ref.extractall()  # Extracts into the current working directory


from transformers import DistilBertForSequenceClassification

# Path to the folder of the extracted model
model_path = "bert-toxic-comment-classification"

# Load the fine-tuned model
model = DistilBertForSequenceClassification.from_pretrained(model_path)


import torch

# Path to the saved file
test_encodings = torch.load("tokenized_test_data.pt")
print("Test encodings loaded successfully!")


from torch.utils.data import DataLoader, TensorDataset

# Configure the device
device = 'cuda' if torch.cuda.is_available() else 'cpu'
print(device)

# Create a TensorDataset from the tokenized data
test_dataset = TensorDataset(
    test_encodings['input_ids'],
    test_encodings['attention_mask']
)

# Create a DataLoader to process the data in batches
test_loader = DataLoader(test_dataset, batch_size=1024, shuffle=False)

from tqdm import tqdm

# Generate predictions in batches with a progress bar
all_predictions = []
model.to(device)  # Move the model to the device
model.eval()
with torch.no_grad():
    for batch in tqdm(test_loader, desc="Generating Predictions"):
        # Move the batch tensors to the device
        input_ids, attention_mask = batch[0].to(device), batch[1].to(device)
        outputs = model(
            input_ids=input_ids,
            attention_mask=attention_mask
        )
        all_predictions.append(torch.sigmoid(outputs.logits).cpu().numpy())

# Combine the results from all batches
import numpy as np
predictions = np.concatenate(all_predictions, axis=0)


import pandas as pd

# Assuming the predictions are a numpy array
predictions_df = pd.DataFrame(predictions, columns=['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate'])
predictions_df.to_csv("predictions_distillbert.csv", index=False)
torch.save(predictions, "predictions_distillbert.pt")

print("Predictions saved")


X_test = joblib.load("X_test.joblib")
y_test = joblib.load("y_test.joblib")
model_logit = joblib.load("logistic_regression_multilabel.joblib")

y_pred_proba = model_logit.predict_proba(X_test)  # Probabilities for each class
y_pred_proba


import pandas as pd
predictions_logit_df = pd.DataFrame(y_pred_proba, columns=['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate'])
predictions_logit_df.to_csv("predictions_logistic_regression.csv", index=False)
print("Predictions saved to predictions_logistic_regression.csv")
